detach the main component as preparation for router introduction.
